Undo CalyxOS 4.7.6 platform package suspensions
Remove any system-level package suspensions for users coming from
CalyxOS 4.7.5 or CalyxOS 4.7.6, the latter of which introduced a bug
preventing packages from being unsuspended as a result of setting
their suspending package to the platform package ("android"). This
change simply removes any suspensions with the platform package as
the suspending package, which in practice should *only* affect apps
that were suspended in 4.7.6, because CalyxOS does not support
device owner use cases.

Users who are not coming from CalyxOS 4.7.5 or 4.7.6 will not have
their suspensions affected at all. Unfortunately, not enough info
is available to limit this to affecting 4.7.6.

Change-Id: I71d5af0725ad12e4704fa795df2aed8333a5e0dc