Don't botch migration if cleartext policy fails
The cleartext network policy shouldn't fail to be migrated, but if it
does for some reason, now we log the error and also set the transaction
as successful before moving on. Previously, because we did not set the
transaction as successful, all database changes made going forward
(including version number) would have been reverted at the end, thus
repeated on every boot.

Change-Id: I8cd5dd949ff34e3c0aaf3f01bf9c314dc91ce926